Make runps-file to use runps-simple & script.ps1
Works much better. Removes dependency on `--file` switch in our
executable (which is not a PowerShell thing) and instead uses the
canonical method of launching a script/executable in PowerShell (the
ampersand). Coincidentally, this now makes the output not depend on a
TTY, and so `monad-run make hashbang-tests` now works as expected.