Fix xUnit tests for `netstandardapp1.5`
With the switch to `netstandardapp1.5` framework for the host, the xUnit
test runner broke. For `dotnet test` to work with the .NET Core xUnit
runner, it must continue to target `dnxcore50`. Fortunately, we can
import our project's framework, and make everything work again.

Also improved the `xunit.sh` script to check for dependencies.