Add Start-PSPackage -Type switch to build any output
If not specified, "deb" is used as default on Linux. This should be
updated to check the distribution and required package build
dependencies.

Dependency names switched per package type, since specific package names
are different. Should check if this causes an issue on OS X when the
names are empty.