Upgrade CLI to build 1718
Per dotnet/cli#1783, this should resolve PowerShell/PowerShell#638.

Unfortunately, per dotnet/cli#1839, the Ubuntu package feed is
out-of-date. So now Travis and AppVeyor aren't quite in sync. This is
not great, but the Ubuntu version of CLI does not have the same
regression that the Windows version had.

The readme has been updated to note how to install an actually
up-to-date copy on Linux.