Add VS Code launch and task files
This enables building PowerShell through VS Code. The build task
launches the installed version of PowerShell and runs `Start-PSBuild`.
This enables a `launch` debug task to immediately debug the `powershell`
process, without having to attach to an external process.

Note that it launches the built version of PowerShell as `powershell -c
'exit'`. This should be modified to suit a user's needs.