Remove conflicting aliases for Linux and OS X
The aliases which conflict with Linux native commands were consolidated
and removed with the pre-processor definition `LINUX`. Additionally,
several "conflicts" are purposefully kept and likewise consolidated.
They are kept because Bash itself stomps on the native commands with its
own built-ins, and so we're adopting the same behavior.

The `clear` and `more` aliases are edge cases where the functions they
call are already fixed for Linux, and so we keep these too.

Resolves #567.