Fix cryptic Pester failure
When the Pester tests were invoked over PSRP and the environment was
more pristine than a local client, the lack of a "foo" variable in some
scope caused the Pester test suite to emit an error at the end, failing
the PSRP build.

Removing the Clear-Variable directive eliminates this error, and is not
necessary for the Pester test itself.