Update XDG BDS tests to specifications
This renames the tests and updates for better requirement
specifications. The profile, modules, history, and startup cache data
are tested for not breaking Windows behavior, new Linux behavior, and
respecting set variables on launch. This also corrects some erroneous
assumptions I had previously made.

The ModuleAnalysisCache cannot be tested, and the startup cache data
might not be testable on Windows.