Improve Dockerfile, reducing image size and build time
Dockerfile instructions each create a filesystem layer, and the current
formulation leaves both a package cache and the powershell deb file to
bulk out the image. The changes here mean that neither are saved in the
final image. This shaves about 80MB from the final image in my testing.

I've also added a dockerignore file. This ignores everything from the
current directory, which speeds up build time by not adding all the
source files from the top level directory to the build context.