grammar: TransactedRegistry stub in System.Management.Automation CoreCLR
PowerShell has some stubs to enable code to almost sort of work with things.
Some of the stubs try to say "Do not use me".

This commit corrects the grammar of the message.

Previously the message tried to apply a past tense form to `hit`,
but the past tense of `hit` is `hit`, and while breakpoints can
be hit, it's better to talk to the user about what should be done/
should not be done.