Allow empty string names for PSObjects #1755
In issue #1755 we determined that you could not create a PSObject with a
key name of an empty string. This is supported in JSON and in
Hashtables.

Simple change was to lower evaluations of name field from
'IsNullOrEmpty' to '== null'.

Simple test to confirm is:
[pscustomobject]@{'' = 'Test'}

Simple JSON to test which results in ths same error:
{
"": "Test"
}

After these changes, both of these cases work and I am able to
succesfully Convert project.lock.json from JSON and back to JSON with no
data loss.