Add UTF-8 BOM to files that will need it
Although the BOM "is neither required nor recommended" for UTF-8 files
by the Unicode standard, it is still necessary for some tools to process
these files correctly now that they contain the © (copyright) character.
In the absence of a BOM, some tools still default to Latin-1 instead of
UTF-8; until such tools are changed to default to a Unicode encoding,
putting a BOM in front of UTF-8 files remains an unfortunate necessity.