Fix here string '`$a=@`"`n'`"'`n`"@'
1. Fix tokenizer
The tokenizer do multiple scans the script line to get tokens. Before
the fix the tokenizer on the first pass examined that string as
double-quoted (Expandable) string not as here string, figured the
average double quotation mark as a closing and then starting with the
single quotation mark continued processing the line as single-quoted
string which had no closing single quotation mark.
The fix stop the first scan pass after getting '=' (assume
assignment-expression '$a=' for next pass).

2. Add test.