Create a semaphore file to enable or disable console host telemetry
We need a way to allow the user to opt out of telemetry and with the config
RFC still in flight, we don't have a centralized way to do that. This implements
a semaphore file which when present will produce telemetry sends
from the console host when it starts up. In order to disable telemetry, simply
delete the file. Setup will need to be changed to handle this.