Added functionality to load Windows Powershell Modules
A lot of binary modules written for Windows Powershell (Full cLR) use GAC for their assemblies.
CoreCLR does not have a GAC and hence fails to locate the assemblies.

Added functionality to search the well known location of GAC on Windows to load the assemblies.
This functionality can be opted-in by adding a key-value to PowershellProperties.json file at $PSHOME.
The key-value pair for enabling GAC loading is: "AllowGacLoading" : "true"