Remove fixed width Unicode escape sequence tokenizing.
The variable width form will require an update to the EditorSyntax so editors don't show "unexpected token" errors when this form is used in a command or variable name.

I will do that after this PR has been accepted unless the team prefer I submit a PR to that project before this PR is accepted.  The same goes for the doc update to the about_Escape_Characters help topic.