Adopt C# approach to Unicode sequence lexing and remove IsSurrogatePair() method.
The C# approach is to have the esc sequence and Unicode esc lexing return a char but also have an out char surrogateChar (lowSurrogate).

This perhaps makes the Backtick() call sites a bit clearer than the use of the IsSurrogateChar() method.

Also change from using s_invalidChar as 0xfffd (replacement char typically used when rendering an invalid byte in the current byte stream e.g UTF-8 reading 0x66, 0xFC, 0x72 - the 0xFC is not valid in this context and an editor might use char 0xFFFD to indicate the character was not understood.  The C# parser returns 0xFFFF (Char.MaxValue) when there is an error and it needs to return something for the return value.

See what you think. I can always revert this change.