Get Pester fixes into PowerShell
There are two fixes in psl-pester:
Add Context information to the pester log which will reduce the name collisions with test names
Update test teardown so exceptions thrown in AfterAll will not leave pester in a bad state.
When this happens, Pester fails all subsequent tests with the error 'Already in Describe'.